Update Image serizlier and testing

+++ spec/poros/weather_spec.rb
@@ -16,7 +16,7 @@ it 'Returns current forecast' do
   expect(current.sunset).to be_a(Time)
   expect(current.feels_like).to be_a(Float)
   expect(current.humidity).to be_a(Integer)
-  expect(current.uvi).to be_a(Integer)
+  expect(current.uvi).to be_a(Float)
   expect(current.visibility).to be_a(Integer)
   
   daily = DailyWeather.new(weather[:daily][0])
